refactoring and renaming methods to prepare for multiple type of process noises
in ParameterWrapper the method setDerivativePriorsEnabled(bool) is no longer available
and it has been substituted with the more general setProcessModelType(ProcessTypes t)

now the management of process model edges is done at the ParameterVerticesManager level.

Now a condition is enforced: new parameter vertices are added forward or backwards with
respect to the existing ones, and NOT in the middle.

Fixed the IMUHandlerFusionTest that was diverging